+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # dartlang plugin changelog
 
 ## unreleased
+- support running Dartino apps on Nucleo board and on an emulated device
 - added support for fast re-start of running Flutter apps
 
 ## 0.6.25